requirement.txt
discord
pytube
ffmpeg
mega.py
dotenv
